+++ R/speedyseq_functions.R
@@ -40,6 +40,10 @@
 #' `group` values. Ignored if `x` has (or is) a phylogenetic tree.
 #' @param tax_adjust 0: no adjustment; 1: phyloseq-compatible adjustment; 2:
 #' conservative adjustment
+#' @param rank_propagation Logical, default TRUE specifying whether to propagate bad ranks
+#' on the right. If FALSE, bad ranks are not propagated to lower ranks. It is 
+#' mainly useful when working with taxonomic tables with informations beyond 
+#' strict hierarchical ranks (e.g. Traits, Functional annotations, etc.).
 #' @export
 #' @return A new phyloseq-class, otu_table, tax_table, XStringset or
 #'   sam_data object depending on the class of the x param
@@ -56,7 +60,8 @@ setGeneric(
   function(x,
            group,
            reorder = FALSE,
-           tax_adjust = 1L) {
+           tax_adjust = 1L,
+          rank_propagation=TRUE) {
     standardGeneric("merge_taxa_vec")
   }
 )
@@ -64,7 +69,8 @@ setGeneric(
 #' @rdname merge_taxa_vec
 setMethod(
   "merge_taxa_vec", "phyloseq",
-  function(x, group, reorder = FALSE, tax_adjust = 1L) {
+  function(x, group, reorder = FALSE, tax_adjust = 1L,
+          rank_propagation=TRUE) {
     stopifnot(ntaxa(x) == length(group))
     stopifnot(tax_adjust %in% c(0L, 1L, 2L))
     # Warn the user if an impossible reordering is requested
@@ -84,7 +90,8 @@ setMethod(
     if (!is.null(x@tax_table) & tax_adjust != 0) {
       tax <- merge_taxa_vec(tax_table(x), group,
         tax_adjust = tax_adjust,
-        reorder = reorder
+        reorder = reorder,
+        rank_propagation=rank_propagation
       )
       # Taxa in `tax` are in same order as in `otu` but are named by first in
       # group instead of max and so need to be renamed
@@ -106,7 +113,8 @@ setMethod(
 #' @rdname merge_taxa_vec
 setMethod(
   "merge_taxa_vec", "otu_table",
-  function(x, group, reorder = FALSE) {
+  function(x, group, reorder = FALSE,
+          rank_propagation=TRUE) {
     stopifnot(ntaxa(x) == length(group))
     # Work with taxa as rows, and remember to flip back at end if needed
     needs_flip <- !taxa_are_rows(x)
@@ -149,7 +157,8 @@ setMethod(
 #' @rdname merge_taxa_vec
 setMethod(
   "merge_taxa_vec", "taxonomyTable",
-  function(x, group, reorder = FALSE, tax_adjust = 1L) {
+  function(x, group, reorder = FALSE, tax_adjust = 1L,
+          rank_propagation=TRUE) {
     stopifnot(ntaxa(x) == length(group))
     # Temporary stopgap to avoid hidden errors if internal variable names are
     # in the tax table
@@ -201,21 +210,21 @@ setMethod(
     reduced_by_group <- reduced_by_group %>%
       select(-.group) %>%
       tibble::column_to_rownames(".taxon")
-    # If only one tax rank, just convert bad_string -> NA; else, need to
+
+    # If rank_propagation is FALSE, just convert bad_string -> NA
     # propagate bad ranks downwards and convert to NAs
-    if (identical(length(rank_names(x)), 1L)) {
-      reduced[[1]] <- reduced[[1]] %>%
-        {
-          ifelse(. == bad_string, NA_character_, .)
-        }
-      reduced %>%
-        as("matrix") %>%
-        tax_table()
-    } else {
-      reduced %>%
+
+    if (rank_propagation){
+      reduced_by_group %>%
         apply(1, bad_flush_right, bad = bad_string, na_bad = na_bad, k = k) %>%
         t() %>%
         tax_table()
+    }
+    else {
+            reduced_by_group |>
+        mutate(across(everything(), ~ ifelse(str_detect(.x, bad_string), NA_character_, .x))) %>%
+        as("matrix") %>%
+        tax_table()   
     }
   }
 )